+++ src/main/java/com/munievents/munievents/repository/PhotosRepository.java
@@ -5,7 +5,7 @@
  import org.springframework.stereotype.Repository;
 
  @Repository
- public interface PhotosRepository  extends JpaRepository<Photos, Long> {
+ public interface PhotosRepository  extends JpaRepository<Photos, Integer> {
 
  }
